+++ src/main/java/algorithms/project/benchmark/Discus.java
@@ -3,8 +3,8 @@ package algorithms.project.benchmark;
 public class Discus implements Benchmark {
 
     @Override
-    public float benchmark(Float... params) {
-        float fitness = 1000000 * params[0] * params[0];
+    public Double benchmark(Double... params) {
+        double fitness = 1000000 * params[0] * params[0];
 
         for (int i = 1; i < params.length; i++) {
             fitness += params[i] * params[i];
